requirements.txt
opencv-python~=4.5.5.62
matplotlib~=3.5.0
numpy~=1.21.2
scipy~=1.7.3
albumentations~=1.0.3
torchvision~=0.11.3
pillow~=8.4.0
requests~=2.24.0
six~=1.16.0

-f https://download.pytorch.org/whl/torch_stable.html
pytorch~=1.10.2